tests/wpt/web-platform-tests/css/css-grid/grid-model/grid-item-accepts-first-letter-001.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Grid Layout Test: '::first-letter' is valid in grid items</title>
<link rel="author" title="Manuel Rego Casasnovas" href="mailto:rego@igalia.com">
<link rel="help" href="http://www.w3.org/TR/css-grid-1/#grid-containers">
<link rel="help" href="http://www.w3.org/TR/css3-selectors/#first-letter">
<meta name="assert" content="The test checks that grid items accept first-letter pseudo-element.">
<link href="/css/support/grid.css" rel="stylesheet">
<style>
  .item::first-letter { line-height: 100px; }
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>

<body onload="checkLayout('.grid,.inline-grid');">

<div id="log"></div>

<div class="grid">
  <div class="item" data-expected-height=100>The first item.</div>
  <div class="item" data-expected-height=100>The second item.</div>
</div>

<div class="inline-grid">
  <div class="item" data-expected-height=100>The first item.</div>
  <div class="item" data-expected-height=100>The second item.</div>
</div>

</body>
